refactor(cockpit): shared LifecycleStateBadge for all operating_model families (DAT-465)
declared/grounded/executed is substrate-generic (current_lifecycle_artifacts),
so the state badge moves to widgets/lifecycle-badges.tsx (rule 13) — cycle and
metric reuse it. validation-badges.tsx keeps only the validation-specific
pass/fail verdict; the two validation widgets import the shared badge.
